SDT-586 - Add express server
