+++ README.md
@@ -32,7 +32,7 @@ Using the command line tool (CLI), you can create a service and add screenshots 
 1. Open your terminal
 2. Navigate to the root directory
 3. If this is the first time you've run the tool you'll have to run `npm install`
-4. Type `npm start`
+4. Type `npm run add`
 5. Follow the prompts
 
 You’ll be given a list of the services that are already in the app and the option to add a new one.
