+++ index.html
@@ -30,17 +30,13 @@
   <h2>The services </h2>
 
   <ul class="exemplar-list">
-    <li><a href="service/tax-account/index.html">Your tax account</a></li>
+    <li><a href="service/your-tax-account/index.html">Your tax account</a></li>
     <li><a href="service/agent-services/index.html">Agent services</a></li>
-    <li><a href="service/stamp-duty-land-tax-calculator/index.html">Stamp duty land tax</a></li>
+    <li><a href="service/stamp-duty-land-tax-calculator/index.html">Stamp duty land tax calculator</a></li>
     <li><a href="service/marriage-allowance/index.html">Marriage Allowance</a></li>
-    <li><a href="service/tax-calculation/index.html">Tax Calculation  Other</a></li>
+    <li><a href="service/tax-calculation/index.html">Tax Calculation Other</a></li>
 </ul>
 </div>
-
-
-
-
