+++ index.js
@@ -11,7 +11,6 @@ var createService = require('./lib/create-service')
 var validationUtil = require('./lib/utils/validation')
 
 var servicesDir = path.join(__dirname, 'service')
-
 var listServices = function () {
   return getServices(servicesDir)
     .then(function (services) {
