+++ lib/update-page.js
@@ -11,20 +11,23 @@ var updatePage = function (indexPage, serviceName, serviceDir) {
 
       var rootDir = path.resolve(__dirname, '..')
       var url = path.join(path.relative(rootDir, serviceDir), 'index.html')
-
       var $ = cheerio.load(contents.toString())
 
       var servicesList = $('.exemplar-list li').map(function (i, el) {
         return $(el).text()
       }).get()
 
-      if (servicesList.includes(serviceName)) {
+      var serviceExists = servicesList.find(function (service) {
+        return service.toUpperCase() === serviceName.toUpperCase()
+      })
+
+      if (serviceExists) {
         resolve(indexPage)
       } else {
         var $li = $('<li></li>')
         var $a = $('<a></a>').attr('href', url).text(serviceName)
 
-        $('.exemplar-list').append('\t', $li.append($a), '\n')
+        $('.exemplar-list').append('\t\t', $li.append($a), '\n')
 
         fs.writeFile(indexPage, $.html(), function (err) {
           if (err) {
