--- service/tax-account/data.js
@@ -1,113 +1,0 @@
-var data = {
-  "service": "17. Your tax account",
-  "last-updated": "1 August 2014",
-  "userjourneys": [
-  {
-    "title": "Public Beta (14 August 2014)",
-    "path": [{
-      "caption": "Sign in screen",
-      "imgref": "images/web-qa-2014-08-14/00-sign-in.png"
-    },{
-      "caption": "Account overview",
-      "imgref": "images/web-qa-2014-08-14/01-account-overview.png"
-    },{
-      "caption": "Messages inbox",
-      "imgref": "images/web-qa-2014-08-14/02b-messages.png"
-    },{
-      "caption": "Messages inbox (empty)",
-      "imgref": "images/web-qa-2014-08-14/02a-messages-empty.png"
-    },{
-      "caption": "Message detail",
-      "imgref": "images/web-qa-2014-08-14/02c-message-detail.png"
-    },{
-      "caption": "Your details",
-      "imgref": "images/web-qa-2014-08-14/03-your-details.png"
-    },{
-      "caption": "Opt in screen",
-      "imgref": "images/web-qa-2014-08-14/04-email-opt-in.png"
-    },{
-      "caption": "Contact HMRC",
-      "imgref": "images/web-qa-2014-08-14/05-contact.png"
-    },{
-      "caption": "Feedback",
-      "imgref": "images/web-qa-2014-08-14/06-feedback.png"
-    },{
-      "caption": "Exit survey",
-      "imgref": "images/web-qa-2014-08-14/08-survey.png"
-    }]
-  },{
-    "title": "Tax account prototype (30 July 2014)",
-    "path": [{
-      "caption": "Opt in screen",
-      "imgref": "images/prototype-2014-07-30/01-self-assessment-opt-in.png"
-    },{
-      "caption": "Opt in screen",
-      "imgref": "images/prototype-2014-07-30/01a-self-assessment-opt-in.png",
-      "note": "Email address entry revealed"
-    },{
-      "caption": "To do list",
-      "imgref": "images/prototype-2014-07-30/02-todo-list.png"
-    },{
-      "caption": "Your taxes overview",
-      "imgref": "images/prototype-2014-07-30/03-your-taxes.png"
-    },{
-      "caption": "Self Assessment detail",
-      "imgref": "images/prototype-2014-07-30/03a-self-assessment.png"
-    },{
-      "caption": "Messages inbox",
-      "imgref": "images/prototype-2014-07-30/04-messages-inbox.png"
-    },{
-      "caption": "Message detail",
-      "imgref": "images/prototype-2014-07-30/04a-message-detail.png"
-    },{
-      "caption": "Your details",
-      "imgref": "images/prototype-2014-07-30/05-your-details.png"
-    }]
-  },{
-    "title": "Tax account prototype",
-    "path": [{
-      "caption": "Sign in",
-      "imgref": "images/prototype/9.sign-in.png"
-    },{
-      "caption": "Account overview",
-      "imgref": "images/prototype/0.account-overview.png"
-    },{
-      "caption": "Account overview",
-      "imgref": "images/prototype/8.returns-up-to-date.png",
-      "note": "Self Assessment returns up to date"
-    },{
-      "caption": "Messages inbox",
-      "imgref": "images/prototype/1.messages.png"
-    },{
-      "caption": "Message detail",
-      "imgref": "images/prototype/2.message-pay.png"
-    },{
-      "caption": "To do",
-      "imgref": "images/prototype/4.to-do.png"
-    },{
-      "caption": "Your taxes",
-      "imgref": "images/prototype/5.your-taxes.png"
-    },{
-      "caption": "Messages",
-      "imgref": "images/prototype/6.messages.png"
-    },{
-      "caption": "Important dates",
-      "imgref": "images/prototype/7.dates.png"
-    },{
-      "caption": "Pay Self Assessment",
-      "imgref": "images/prototype/10.pay.png"
-    },{
-      "caption": "Pay by Direct Debit",
-      "imgref": "images/prototype/11.pay-direct-debit.png"
-    },{
-      "caption": "Pay by credit or debit card",
-      "imgref": "images/prototype/12.credit-debit-card.png"
-    },{
-      "caption": "Pay by bank transfer",
-      "imgref": "images/prototype/13.bank-account-transfer.png"
-    },{
-      "caption": "Sign out",
-      "imgref": "images/prototype/3.sign-out.png"
-    }]
-  }]
-}
